Massive Cleanup
Configures Resharper, StyleCop and the SonarQube analyzers, cleaned up the solution quite a bit, tried to align the code style in a way that seems to suit most team members and aligns closely to the default stylecop rules.

Added quite a bit of documentation to commonly used classes.

Cleaned up or simplified a lot of complex conditionals.