Unified solution for TFS 2013 and 2015
Refactored solution to support a (Debug|Release)-(2013|2015), an MsBuild Property called TfsVersion to load references for each version and a build constant to use in the sources (TFS2013|TFS2015).

Build and tests for 2015 succeed. 2013 has not been tested in this commit and is likely to fail.

Solution depends on Visual Studio 2015 to build (uses new C#6 langage features).